transformed plotting into a function
the heatmap code was transformed into a function to handled repeated
clicks on the options. The call to the php was logged to console
for debugging.

The next thing to fix is how the data are written to the MySQL DB.
The column types need to be specified in order to handle the dates.